Crate substrate_finality_grandpa

Integration of the GRANDPA finality gadget into substrate.

This crate is unstable and the API and usage may change.

This crate provides a long-running future that produces finality notifications.

Usage

First, create a block-import wrapper with the block_import function. The GRANDPA worker needs to be linked together with this block import object, so a LinkHalf is returned as well. All blocks imported (from network or consensus or otherwise) must pass through this wrapper, otherwise consensus is likely to break in unexpected ways.

Next, use the LinkHalf and a local configuration to run_grandpa_voter. This requires a Network implementation. The returned future should be driven to completion and will finalize blocks in the background.

Changing authority sets

The rough idea behind changing authority sets in GRANDPA is that at some point, we obtain agreement for some maximum block height that the current set can finalize, and once a block with that height is finalized the next set will pick up finalization from there.

Technically speaking, this would be implemented as a voting rule which says, "if there is a signal for a change in N blocks in block B, only vote on chains with length NUM(B) + N if they contain B". This conditional-inclusion logic is complex to compute because it requires looking arbitrarily far back in the chain.

Instead, we keep track of a list of all signals we've seen so far (across all forks), sorted ascending by the block number they would be applied at. We never vote on chains with number higher than the earliest handoff block number (this is num(signal) + N). When finalizing a block, we either apply or prune any signaled changes based on whether the signaling block is included in the newly-finalized chain.

Structs

Config
Configuration for the GRANDPA service.

FinalityProofProvider
Finality proof provider for serving network requests.

GrandpaParams
Parameters used to run Grandpa.

LinkHalf

ScheduledChange
A scheduled change of authority set.

Enums

Error
Errors that can occur while voting in GRANDPA.

Traits

BlockStatus
Something which can determine if a block is known.

Network
A handle to the network. This is generally implemented by providing some handle to a gossip service or similar.

Functions

block_import
Make block importer and link half necessary to tie the background voter to it.

light_block_import
Create light block importer.

run_grandpa	Deprecated

run_grandpa_observer
Run a GRANDPA observer as a task, the observer will finalize blocks only by listening for and validating GRANDPA commits instead of following the full protocol. Provide configuration and a link to a block import worker that has already been instantiated with block_import.

run_grandpa_voter
Run a GRANDPA voter as a task. Provide configuration and a link to a block import worker that has already been instantiated with block_import.

setup_disabled_grandpa
When GRANDPA is not initialized we still need to register the finality tracker inherent provider which might be expected by the runtime for block authoring. Additionally, we register a gossip message validator that discards all GRANDPA messages (otherwise, we end up banning nodes that send us a Neighbor message, since there is no registered gossip validator for the engine id defined in the message.)

Type Definitions

AuthorityId
Identity of a Grandpa authority.

CatchUp
A catch up message for this chain's block type.

Commit
A commit message for this chain's block type.

CompactCommit
A compact commit message for this chain's block type.

Message
A GRANDPA message for a substrate chain.

Precommit
A precommit message for this chain's block type.

Prevote
A prevote message for this chain's block type.

PrimaryPropose
A primary propose message for this chain's block type.

SignedMessage
A signed message.