Function substrate_finality_grandpa::run_grandpa_voter

pub fn run_grandpa_voter<B, E, Block: BlockT<Hash = H256>, N, RA, SC, X>(
    grandpa_params: GrandpaParams<B, E, Block, N, RA, SC, X>
) -> Result<impl Future<Item = (), Error = ()> + Send + 'static> where
    Block::Hash: Ord,
    B: Backend<Block, Blake2Hasher> + 'static,
    E: CallExecutor<Block, Blake2Hasher> + Send + Sync + 'static,
    N: Network<Block> + Send + Sync + 'static,
    N::In: Send + 'static,
    SC: SelectChain<Block> + 'static,
    NumberFor<Block>: BlockNumberOps,
    DigestFor<Block>: Encode,
    RA: Send + Sync + 'static,
    X: Future<Item = (), Error = ()> + Clone + Send + 'static, 

Run a GRANDPA voter as a task. Provide configuration and a link to a block import worker that has already been instantiated with block_import.