Trait substrate_finality_grandpa::BlockStatus

pub trait BlockStatus<Block: BlockT> {
    fn block_number(
        &self, 
        hash: Block::Hash
    ) -> Result<Option<NumberFor<Block>>, Error>;
}

Something which can determine if a block is known.

Required methods

fn block_number(
&self,
hash: Block::Hash
) -> Result<Option<NumberFor<Block>>, Error>

Return Ok(Some(number)) or Ok(None) depending on whether the block is definitely known and has been imported. If an unexpected error occurs, return that.

Loading content...