Trait substrate_finality_grandpa::Network

pub trait Network<Block: BlockT>: Clone {
    type In: Stream<Item = Vec<u8>, Error = ()>;
    fn messages_for(&self, round: u64, set_id: u64) -> Self::In;

    fn send_message(
        &self, 
        round: u64, 
        set_id: u64, 
        message: Vec<u8>, 
        force: bool
    );

    fn drop_round_messages(&self, round: u64, set_id: u64);

    fn drop_set_messages(&self, set_id: u64);

    fn commit_messages(&self, set_id: u64) -> Self::In;

    fn send_commit(
        &self, 
        round: u64, 
        set_id: u64, 
        message: Vec<u8>, 
        force: bool
    );

    fn announce(&self, round: u64, set_id: u64, block: Block::Hash);
}

A handle to the network. This is generally implemented by providing some handle to a gossip service or similar.

Intended to be a lightweight handle such as an Arc.

Associated Types

type In: Stream<Item = Vec<u8>, Error = ()>

A stream of input messages for a topic.

Loading content...
Required methods

fn messages_for(&self, round: u64, set_id: u64) -> Self::In

Get a stream of messages for a specific round. This stream should never logically conclude.

fn send_message(&self, round: u64, set_id: u64, message: Vec<u8>, force: bool)

Send a message at a specific round out.

fn drop_round_messages(&self, round: u64, set_id: u64)

Clean up messages for a round.

fn drop_set_messages(&self, set_id: u64)

Clean up messages for a given authority set id (e.g. commit messages).

fn commit_messages(&self, set_id: u64) -> Self::In

Get a stream of commit messages for a specific set-id. This stream should never logically conclude.

fn send_commit(&self, round: u64, set_id: u64, message: Vec<u8>, force: bool)

Send message over the commit channel.

fn announce(&self, round: u64, set_id: u64, block: Block::Hash)

Inform peers that a block with given hash should be downloaded.

Loading content...