Trait substrate_finality_grandpa::Network

pub trait Network<Block: BlockT>: Clone + Send + 'static {
    type In: Stream<Item = TopicNotification, Error = ()>;
    fn messages_for(&self, topic: Block::Hash) -> Self::In;

    fn register_validator(&self, validator: Arc<dyn Validator<Block>>);

    fn gossip_message(&self, topic: Block::Hash, data: Vec<u8>, force: bool);

    fn register_gossip_message(&self, topic: Block::Hash, data: Vec<u8>);

    fn send_message(&self, who: Vec<PeerId>, data: Vec<u8>);

    fn report(&self, who: PeerId, cost_benefit: i32);

    fn announce(&self, block: Block::Hash);
}

A handle to the network. This is generally implemented by providing some handle to a gossip service or similar.

Intended to be a lightweight handle such as an Arc.

Associated Types

type In: Stream<Item = TopicNotification, Error = ()>

A stream of input messages for a topic.

Loading content...
Required methods

fn messages_for(&self, topic: Block::Hash) -> Self::In

Get a stream of messages for a specific gossip topic.

fn register_validator(&self, validator: Arc<dyn Validator<Block>>)

Register a gossip validator.

fn gossip_message(&self, topic: Block::Hash, data: Vec<u8>, force: bool)

Gossip a message out to all connected peers.

Force causes it to be sent to all peers, even if they've seen it already. Only should be used in case of consensus stall.

fn register_gossip_message(&self, topic: Block::Hash, data: Vec<u8>)

Register a message with the gossip service, it isn't broadcast right away to any peers, but may be sent to new peers joining or when asked to broadcast the topic. Useful to register previous messages on node startup.

fn send_message(&self, who: Vec<PeerId>, data: Vec<u8>)

Send a message to a bunch of specific peers, even if they've seen it already.

fn report(&self, who: PeerId, cost_benefit: i32)

Report a peer's cost or benefit after some action.

fn announce(&self, block: Block::Hash)

Inform peers that a block with given hash should be downloaded.

Loading content...